Allow for conservative comparison to MICE
In our non-MICE method, we get to use the y values in the training set when deciding how to handle missingness. In the naive semi-supervised approach from before, we ran mice on the covariates of the train and test sets, but no y values whatsovever. An extremely MICE-favouring comparison would be to allow MICE to see the train *and test* labels, and see if we can still do better. It's not clear how to allow MICE to use train labels but not test labels, at least out-of-the-box.