Update experiment pipeline for fixed lambda
now includes objective and accuracy, both with error bars. Notebook file should be set up for fairly easy running of experiments for different datasets now. We may want to also add functionality to plot (a) sparsity vs lambda and (b) performance on different subsets of the data (i.e. missing and not).